Internalize filetype detection.
